remove team id from worker resource check sessions
In order to make check containers global, we removed the team
id field from the worker resource config check sessions table. As a
result, check containers are not associated to a team id which made it
not make sense to have the create/find containers to live on the team
object. The create/find containers methods are moved to the worker
object and the FindWorkerForContainerByOwner is moved to the worker
factory. (The find container methods that are only used by hijack is
left on the team object because check containers are no longer
hijackable.) We want check containers to be no longer hijackable because
now that check containers are global, we wouldn't want teams to be
able to mess with a container that is shared by all teams.

concourse/concourse#2386

Signed-off-by: Saman Alvi <salvi@pivotal.io>